feat: Restructure project into modules for better readability and maintainability.
