+++ app/scrapers/__init__.py
@@ -1,0 +1,2 @@
+from .base import ShopItem
+from .alza import AlzaItem
