+++ app/services/__init__.py
@@ -1,0 +1,2 @@
+from .email_notifier import EmailNotifierInterface
+from .price_tracker import PriceTracker
