+++ requirements.txt
@@ -1,0 +1,1 @@
+selenium>=4.27.1
